Refactoring agentList implementation
