+++ frontend/app/[locale]/agents/components/AgentManageComp.tsx
@@ -4,73 +4,34 @@ import { useTranslation } from "react-i18next";
 import { App, Row, Col, Flex, Tooltip, Badge, Divider } from "antd";
 import { FileInput, Plus, X } from "lucide-react";
 
-import { Agent } from "@/types/agentConfig";
 import AgentList from "./agentManage/AgentList";
-import { useSaveGuard } from "@/hooks/agent/useSaveGuard";
-import { useCallback } from "react";
+
 import { useAgentConfigStore } from "@/stores/agentConfigStore";
 import { importAgent } from "@/services/agentConfigService";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useAgentList } from "@/hooks/agent/useAgentList";
-import { useAgentInfo } from "@/hooks/agent/useAgentInfo";
 import log from "@/lib/logger";
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import { ImportAgentData } from "@/hooks/useAgentImport";
 import AgentImportWizard from "@/components/agent/AgentImportWizard";
-import { clearAgentNewMark } from "@/services/agentConfigService";
+
 
 export default function AgentManageComp() {
   const { t } = useTranslation("common");
   const { message } = App.useApp();
 
   // Get state from store
-  const currentAgentId = useAgentConfigStore((state) => state.currentAgentId);
-  const hasUnsavedChanges = useAgentConfigStore(
-    (state) => state.hasUnsavedChanges
-  );
   const isCreatingMode = useAgentConfigStore((state) => state.isCreatingMode);
-  const setCurrentAgent = useAgentConfigStore((state) => state.setCurrentAgent);
   const enterCreateMode = useAgentConfigStore((state) => state.enterCreateMode);
   const reset = useAgentConfigStore((state) => state.reset);
-
-  // Unsaved changes guard
-  const checkUnsavedChanges = useSaveGuard();
-
-  // Handle unsaved changes check and agent switching
-  const handleAgentSwitch = useCallback(
-    async (agentDetail: any) => {
-      const canSwitch = await checkUnsavedChanges.saveWithModal();
-      if (canSwitch) {
-        setCurrentAgent(agentDetail);
-      }
-    },
-    [checkUnsavedChanges]
-  );
-
-  const editable = currentAgentId || isCreatingMode;
-
-  // Shared agent list via React Query
-  const { agents: agentList, isLoading: loading, refetch } = useAgentList();
-  const queryClient = useQueryClient();
-
-  // State for selected agent info loading
-  const [selectedAgentId, setSelectedAgentId] = useState<number | null>(null);
 
   // Import wizard state
   const [importWizardVisible, setImportWizardVisible] = useState(false);
   const [importWizardData, setImportWizardData] =
     useState<ImportAgentData | null>(null);
 
-  const {
-    data: agentDetail,
-    isLoading: agentInfoLoading,
-    error: agentInfoError,
-  } = useAgentInfo(selectedAgentId);
-
-  const importAgentMutation = useMutation({
-    mutationFn: (agentData: any) => importAgent(agentData),
-    onSuccess: () => queryClient.invalidateQueries({ queryKey: ["agents"] }),
-  });
+  // Shared agent list via React Query
+  const { agents: agentList, isLoading: loading, refetch } = useAgentList();
 
   // Handle import agent for space view - open wizard instead of direct import
   const handleImportAgent = () => {
@@ -114,63 +75,6 @@ export default function AgentManageComp() {
     };
 
     fileInput.click();
-  };
-
-  // Handle agent detail loading completion
-  useEffect(() => {
-    if (
-      selectedAgentId &&
-      agentDetail &&
-      !agentInfoLoading &&
-      !agentInfoError
-    ) {
-      // Handle agent switch with unsaved changes check
-      handleAgentSwitch(agentDetail);
-      setSelectedAgentId(null);
-    } else if (selectedAgentId && agentInfoError && !agentInfoLoading) {
-      // Handle error
-      log.error("Failed to load agent detail:", agentInfoError);
-      message.error(t("agentConfig.agents.detailsLoadFailed"));
-      setSelectedAgentId(null);
-    }
-  }, [
-    selectedAgentId,
-    agentDetail,
-    agentInfoLoading,
-    agentInfoError,
-    handleAgentSwitch,
-    message,
-    t,
-  ]);
-
-  // Handle select agent
-  const handleSelectAgent = async (agent: Agent) => {
-    // Clear NEW mark when agent is selected for editing (only if marked as new)
-    if (agent.is_new === true) {
-      try {
-        const res = await clearAgentNewMark(agent.id);
-        if (res?.success) {
-          log.warn("Failed to clear NEW mark on select:", res);
-          queryClient.invalidateQueries({ queryKey: ["agents"] });
-        }
-      } catch (err) {
-        log.error("Failed to clear NEW mark on select:", err);
-      }
-    }
-
-    // If already selected, deselect it
-    if (
-      currentAgentId !== null &&
-      String(currentAgentId) === String(agent.id)
-    ) {
-      const canDeselect = await checkUnsavedChanges.saveWithModal();
-      if (canDeselect) {
-        setCurrentAgent(null);
-      }
-      return;
-    }
-    // Set selected agent id to trigger the hook
-    setSelectedAgentId(Number(agent.id));
   };
 
   return (
@@ -282,17 +186,7 @@ export default function AgentManageComp() {
         </Row>
 
         <div className="flex-1 min-h-0">
-          <AgentList
-            agentList={agentList}
-            currentAgentId={currentAgentId}
-            hasUnsavedChanges={hasUnsavedChanges}
-            onSelectAgent={handleSelectAgent}
-            onAgentDeleted={(agentId) => {
-              if (currentAgentId === agentId) {
-                setCurrentAgent(null);
-              }
-            }}
-          />
+          <AgentList agentList={agentList} />
         </div>
       </Flex>
